Made it so the "Drop Dropping Rare Items" feature works when clicking outside of any inventory with the item- not just your player inventory. Drop confirmation & stop dropping rare items now treats all items with a different name as a separate item (fixes all player heads being recognized as 1 item). "RARE" items are now recognized as rare within your inventory, annd "UNCOMMON" items are recognized as rare within your hotbar only.

+++ src/main/java/codes/biscuit/skyblockaddons/tweaker/SkyblockAddonsTransformer.java
@@ -60,7 +60,6 @@ public class SkyblockAddonsTransformer implements IClassTransformer {
         registerTransformer(new MinecraftTransformer());
         registerTransformer(new ItemTransformer());
         registerTransformer(new GuiScreenTransformer());
-        registerTransformer(new GuiInventoryTransformer());
         registerTransformer(new GuiContainerTransformer());
         registerTransformer(new GuiChestTransformer());
         registerTransformer(new GuiNewChatTransformer());
@@ -113,10 +112,12 @@ public class SkyblockAddonsTransformer implements IClassTransformer {
         if (OUTPUT_BYTECODE) {
             try {
                 File file = new File("C:\\Users\\jlroc\\Desktop\\bytecode", transformedName + ".class");
-                file.createNewFile();
-                FileOutputStream os = new FileOutputStream(file);
-                os.write(writer.toByteArray());
-                os.close();
+                if (file.getParentFile().exists()) {
+                    file.createNewFile();
+                    FileOutputStream os = new FileOutputStream(file);
+                    os.write(writer.toByteArray());
+                    os.close();
+                }
             } catch (IOException e) {
                 e.printStackTrace();
             }
